Move get pull requests in build
