+++ lib/controllers/deploy-sites.js
@@ -1,4 +1,4 @@
-const githubServices = require('../services/github');
+const githubServices = require('../../common/services/github');
 const releasesService = require('../services/releases');
 const config = require('../../config');
 const Boom = require('@hapi/boom');
