+++ lib/controllers/slack.js
@@ -1,7 +1,7 @@
 const commands = require('../../common/services/slack/commands');
 const shortcuts = require('../../common/services/slack/shortcuts');
 const viewSubmissions = require('../../common/services/slack/view-submissions');
-const github = require('../services/github');
+const github = require('../../common/services/github');
 const postSlackMessage = require('../../common/services/slack/surfaces/messages/post-message');
 
 function _getDeployStartedMessage(release, appName) {
@@ -94,11 +94,6 @@ module.exports = {
       console.log('This kind of interaction is not yet supported by Pix Bot.');
       return null;
     }
-  },
-
-  async getPullRequests(request) {
-    const label = request.pre.payload.text;
-    return github.getPullRequests(label);
   },
 
   async getChangelogSinceLatestRelease() {
