+++ lib/routes/slack.js
@@ -50,12 +50,6 @@ module.exports = [
   },
   {
     method: 'POST',
-    path: '/slack/commands/pull-requests',
-    handler: slackbotController.getPullRequests,
-    config: slackConfig
-  },
-  {
-    method: 'POST',
     path: '/slack/interactive-endpoint',
     handler: slackbotController.interactiveEndpoint,
     config: slackConfig
