+++ test/unit/lib/services/github_test.js
@@ -2,7 +2,7 @@ const { describe, it } = require('mocha');
 const { expect } = require('chai');
 const { sinon, nock } = require('../../../test-helper');
 const axios = require('axios');
-const githubService = require('../../../../lib/services/github');
+const githubService = require('../../../../common/services/github');
 
 describe('#getPullRequests', function() {
   const items = [
